Move some common models to a core package
