+++ infobip_channels/whatsapp/channel.py
@@ -5,11 +5,12 @@ import requests
 from pydantic import AnyHttpUrl, BaseModel
 from pydantic.error_wrappers import ValidationError
 
-from infobip_channels.http_client import _HttpClient
+from infobip_channels.core.http_client import _HttpClient
+from infobip_channels.core.models import Authentication, GetHeaders, PostHeaders
 from infobip_channels.whatsapp.models.body.audio_message import AudioMessageBody
 from infobip_channels.whatsapp.models.body.buttons_message import ButtonsMessageBody
 from infobip_channels.whatsapp.models.body.contact_message import ContactMessageBody
-from infobip_channels.whatsapp.models.body.core import Authentication, MessageBody
+from infobip_channels.whatsapp.models.body.core import MessageBody
 from infobip_channels.whatsapp.models.body.create_template import CreateTemplate
 from infobip_channels.whatsapp.models.body.document_message import DocumentMessageBody
 from infobip_channels.whatsapp.models.body.image_message import ImageMessageBody
@@ -23,8 +24,6 @@ from infobip_channels.whatsapp.models.body.sticker_message import StickerMessage
 from infobip_channels.whatsapp.models.body.template_message import TemplateMessageBody
 from infobip_channels.whatsapp.models.body.text_message import TextMessageBody
 from infobip_channels.whatsapp.models.body.video_message import VideoMessageBody
-from infobip_channels.whatsapp.models.headers.get import GetHeaders
-from infobip_channels.whatsapp.models.headers.post import PostHeaders
 from infobip_channels.whatsapp.models.path_parameters.core import PathParameter
 from infobip_channels.whatsapp.models.path_parameters.manage_templates import (
     ManageTemplatesPathParameters,
