+++ infobip_channels/whatsapp/models/body/buttons_message.py
@@ -8,11 +8,8 @@ except ImportError:
 
 from pydantic import AnyHttpUrl, conlist, constr, validator
 
-from infobip_channels.whatsapp.models.body.core import (
-    MessageBody,
-    UrlLengthValidatorMixin,
-)
-from infobip_channels.whatsapp.models.core import CamelCaseModel
+from infobip_channels.core.models import CamelCaseModel, UrlLengthValidatorMixin
+from infobip_channels.whatsapp.models.body.core import MessageBody
 
 
 class ButtonTypeEnum(str, Enum):
