+++ infobip_channels/whatsapp/models/body/contact_message.py
@@ -4,8 +4,8 @@ from typing import List, Optional
 
 from pydantic import validator
 
+from infobip_channels.core.models import CamelCaseModel
 from infobip_channels.whatsapp.models.body.core import MessageBody
-from infobip_channels.whatsapp.models.core import CamelCaseModel
 
 
 class AddressTypeEnum(str, Enum):
