+++ infobip_channels/whatsapp/models/body/create_template.py
@@ -8,8 +8,7 @@ except ImportError:
 
 from pydantic import AnyHttpUrl, conlist, constr, validator
 
-from infobip_channels.whatsapp.models.body.core import UrlLengthValidatorMixin
-from infobip_channels.whatsapp.models.core import CamelCaseModel
+from infobip_channels.core.models import CamelCaseModel, UrlLengthValidatorMixin
 
 
 class LanguageEnum(str, Enum):
