+++ infobip_channels/whatsapp/models/body/document_message.py
@@ -2,11 +2,8 @@ from typing import Optional
 
 from pydantic import AnyHttpUrl, constr, validator
 
-from infobip_channels.whatsapp.models.body.core import (
-    MessageBody,
-    UrlLengthValidatorMixin,
-)
-from infobip_channels.whatsapp.models.core import CamelCaseModel
+from infobip_channels.core.models import CamelCaseModel, UrlLengthValidatorMixin
+from infobip_channels.whatsapp.models.body.core import MessageBody
 
 
 class Content(UrlLengthValidatorMixin, CamelCaseModel):
