+++ infobip_channels/whatsapp/models/body/multi_product_message.py
@@ -3,12 +3,12 @@ from typing import List, Optional
 
 from pydantic import conlist, constr, validator
 
+from infobip_channels.core.models import CamelCaseModel
 from infobip_channels.whatsapp.models.body.action_sections import (
     SectionBase,
     SectionTitleValidatorMixin,
 )
 from infobip_channels.whatsapp.models.body.core import MessageBody
-from infobip_channels.whatsapp.models.core import CamelCaseModel
 
 
 class HeaderTypeEnum(str, Enum):
