+++ infobip_channels/whatsapp/models/body/sticker_message.py
@@ -1,10 +1,7 @@
 from pydantic import AnyHttpUrl, validator
 
-from infobip_channels.whatsapp.models.body.core import (
-    MessageBody,
-    UrlLengthValidatorMixin,
-)
-from infobip_channels.whatsapp.models.core import CamelCaseModel
+from infobip_channels.core.models import CamelCaseModel, UrlLengthValidatorMixin
+from infobip_channels.whatsapp.models.body.core import MessageBody
 
 
 class Content(UrlLengthValidatorMixin, CamelCaseModel):
