+++ infobip_channels/whatsapp/models/body/template_message.py
@@ -7,11 +7,8 @@ except ImportError:
 
 from pydantic import AnyHttpUrl, Field, confloat, conlist, constr, validator
 
-from infobip_channels.whatsapp.models.body.core import (
-    MessageBody,
-    UrlLengthValidatorMixin,
-)
-from infobip_channels.whatsapp.models.core import CamelCaseModel
+from infobip_channels.core.models import CamelCaseModel, UrlLengthValidatorMixin
+from infobip_channels.whatsapp.models.body.core import MessageBody
 
 
 class ButtonQuickReply(CamelCaseModel):
