--- infobip_channels/whatsapp/models/core.py
@@ -1,12 +1,0 @@
-from pydantic import BaseModel
-
-
-def to_camel_case(string: str) -> str:
-    output = "".join(word.capitalize() for word in string.split("_"))
-    return output[0].lower() + output[1:]
-
-
-class CamelCaseModel(BaseModel):
-    class Config:
-        alias_generator = to_camel_case
-        allow_population_by_field_name = True
