--- infobip_channels/whatsapp/models/headers/core.py
@@ -1,20 +1,0 @@
-from typing import Optional
-
-from pydantic import BaseModel
-
-
-def to_header_specific_case(string: str) -> str:
-    return "-".join(word.capitalize() for word in string.split("_"))
-
-
-class RequestHeaders(BaseModel):
-    authorization: str
-    accept: Optional[str] = "application/json"
-
-    class Config:
-        alias_generator = to_header_specific_case
-        allow_population_by_field_name = True
-
-    def __init__(self, **data: str) -> None:
-        super().__init__(**data)
-        self.authorization = f"App {self.authorization}"
