--- infobip_channels/whatsapp/models/headers/get.py
@@ -1,5 +1,0 @@
-from infobip_channels.whatsapp.models.headers.core import RequestHeaders
-
-
-class GetHeaders(RequestHeaders):
-    pass
