--- infobip_channels/whatsapp/models/headers/post.py
@@ -1,7 +1,0 @@
-from typing import Optional
-
-from infobip_channels.whatsapp.models.headers.core import RequestHeaders
-
-
-class PostHeaders(RequestHeaders):
-    content_type: Optional[str] = "application/json"
