+++ tests/whatsapp/models/test_core.py
@@ -3,8 +3,9 @@ from pydantic import AnyHttpUrl
 from pydantic.error_wrappers import ValidationError
 from pydantic_factories import ModelFactory
 
+from infobip_channels.core.models import Authentication
 from infobip_channels.whatsapp.channel import WhatsAppChannel
-from infobip_channels.whatsapp.models.body.core import Authentication, MessageBody
+from infobip_channels.whatsapp.models.body.core import MessageBody
 from tests.whatsapp.conftest import get_random_string
